major modifications to buildInputLists.py and submitSkimJobs.py
Details: 

These scripts are still work-in-progress, but they can now function
more or less properly. 
- buildInputLists now just takes n input directory, a list of sample
names, and from that it builds the filelists
- submitSkimJobs.py uses the filelists from buildInputLists, and it
submit the jobs. Jobnames are now matching the sample name, jobs are
submitted to at3 (it used to be at3_xxl), and the output files are
copied to pnfs with names *_skimMu.root and *skimEl.root.
- still to be verified that we obtain correctly the grid proxy (it
seems to work 50% of the times, and I think it was related to env var
and subprocesses)