EventVarInterface: add catch(out_of_range) for event weight container
Details:
despite the checks on the size of the weight container,
sometimes there is an exception thrown for an index out
of range. This only happens for one sample (107704),
and it causes a crash. This fix should bypass the problem,
but it would be nice to undestand what's going on (right
now it's not clear, not even with gdb).