Initial import
Details:
This is the initial import of some code that we're using to understand
whether/how we can apply the Semileptonic b corretion to jets. The
code is a simplified re-implementation of the procedure described in
ATL-COM-PHYS-2008-086. The code is currently working on pre-slim D3PD
files (where the mc-truth jet collection is available). It can produce
basic plots, but it still needs some work.