first version of a script to check the constistency of a JO file
Details:
This script will perform a sanity check of the job option files.
See the comments in the code to get an idea of how it should work.
Run with '-h' to get a list of possible options.
The script is already working, but not all of the possible parameter
values are listed.