EventVariables: fix previous commit
Details:
- during the previous commit (move from .h to .cxx), few new functions
were lost. For this reason I just re-did the migration .h -> .cxx from
scratch.
- move from the .h to the .cxx all the functions that are more than
one single line
- cleanup redundant empty lines
- consistent indentation (please please use consistent indentation -->
use emacs or set your tab to a reasonable number of spaces)
- compilation/linking time much improved: used to be >5min for any
change. Now if you change something in a cxx file, only that file is
re-compiled (few sec) and then everything is linked as it should be.
- all the changes are a re-organization/cleanup only; there is no
change in the implementation, so all the results should be exactly the
same. I cannot check all the JO, but please let me know if you see
something different between this commit and the previous revision.