script to submit all the lists in a given file
Details:
this script can be used to submit the skimming jobs of a set of
filelists. The jobs are submitted after waiting for a given delay, so
that we do not overload pnfs.
Not a smart script, but it might be useful in the future.
Remember to take care of the grid proxy.