add script to merge files within one dataset
Details:
This script is used to merge the files from a given dataset so as to
reduce the number of files per dataset.  Same syntax as 'hadd', but
works with directories rather than files (run without arguments to get
the help message).  By default it does not overwrite, but rather it
skips ouput directories that already have some content.  It does not
perform any kind of checks on the input files (eg. if they have the
same branches etc.).  So far, it has been tested on data samples
(slimmed).