checkGetEntriesSpeed.py: initial import
Details:

Given two directories A and B, this script loops over the common 1st
level subdirectories, and checks that each one of them has trees with
the same total number of entries.
This script is mostly meant to (1) check that when we merge files we
are not loosing events (2) investigate possible speed improvments for
the dataset monitoring script (Javier's html producer).

So far I have tried 3 options: TChain::GetEntries, TTree::GetEntries,
TTree::GetEntriesFast, TTree::GetEntriesFast with python
multiprocessing.  So far it hasn't been a success: all the GetEntries*
perform about the same, and the multiprocessing makes things even
slower. To me, this is a clear indication that the bottleneck is in
fact dcap (and not any TTree access speed). Tomorrow I will implement
some caching in order to avoid all this.