bugfix xsec skimmed W+jets,Z+jets; re-implement part of FillXML
Details:
This bug was due to the fact that the vector of xsec values was filled
without filtering on the files to be accepted or not. MC samples have
two skims (skimEl, skimMu), so the resulting vectors were:
  m_fileNames_tmp[N]
  m_xsec_tmp[2N]
Therefore, the xsec being used for skimmed (W+jets | Z+jets) samples
was wrong.
Also done some other cleanup to make the code more readable.
(Francesco and Davide)