add FcncTruth and NtupleInterface
Details:

classes to determine the truth indices for the particles that are
relevant to FCNC samples. This code compiles and works as standalone,
but it needs some modifications in order to be used within the
Histogrammer+MiniSl (will probably need to drop the NtupleInterface).
Working on it; will add the relevant lines in linkdef and module.mk
when the code is integrated in the Histogrammer.